+++ content_type_field_validations.go
@@ -118,3 +118,61 @@ func (v *FieldValidationSize) UnmarshalJSON(data []byte) error {
 
 	return nil
 }
+
+const (
+	FieldValidationRegexPatternEmail         = `^\w[\w.-]*@([\w-]+\.)+[\w-]+$`
+	FieldValidationRegexPatternURL           = `^(ftp|http|https):\/\/(\w+:{0,1}\w*@)?(\S+)(:[0-9]+)?(\/|\/([\w#!:.?+=&%@!\-\/]))?$`
+	FieldValidationRegexPatternUSDate        = `^(0?[1-9]|[12][0-9]|3[01])[- \/.](0?[1-9]|1[012])[- \/.](19|20)?\d\d$`
+	FieldValidationRegexPatternEuorpeanDate  = `^(0?[1-9]|[12][0-9]|3[01])[- \/.](0?[1-9]|1[012])[- \/.](19|20)?\d\d$`
+	FieldValidationRegexPattern12HourTime    = `^(0?[1-9]|1[012]):[0-5][0-9](:[0-5][0-9])?\s*[aApP][mM]$`
+	FieldValidationRegexPattern24HourTime    = `^(0?[0-9]|1[0-9]|2[0-3]):[0-5][0-9](:[0-5][0-9])?$`
+	FieldValidationRegexPatternUSPhoneNumber = `^\d[ -.]?\(?\d\d\d\)?[ -.]?\d\d\d[ -.]?\d\d\d\d$`
+	FieldValidationRegexPatternUSZipCode     = `^\d{5}$|^\d{5}-\d{4}$}`
+)
+
+// FieldValidationRegex model
+type FieldValidationRegex struct {
+	Pattern      string
+	Flags        string
+	ErrorMessage string
+}
+
+// MarshalJSON for custom json marshaling
+func (v *FieldValidationRegex) MarshalJSON() ([]byte, error) {
+	type regex struct {
+		Pattern string `json:"pattern"`
+		Flags   string `json:"flags"`
+	}
+
+	return json.Marshal(&struct {
+		Regexp  *regex `json:"regexp,omitempty"`
+		Message string `json:"message,omitempty"`
+	}{
+		Regexp:  &regex{v.Pattern, v.Flags},
+		Message: v.ErrorMessage,
+	})
+}
+
+// UnmarshalJSON for custom json unmarshaling
+func (v *FieldValidationRegex) UnmarshalJSON(data []byte) error {
+	payload := map[string]interface{}{}
+	if err := json.Unmarshal(data, &payload); err != nil {
+		return err
+	}
+
+	regex := payload["regexp"].(map[string]interface{})
+
+	if val, ok := regex["pattern"].(string); ok {
+		v.Pattern = val
+	}
+
+	if val, ok := regex["flags"].(string); ok {
+		v.Flags = val
+	}
+
+	if val, ok := regex["message"].(string); ok {
+		v.ErrorMessage = val
+	}
+
+	return nil
+}